Split out Summary builder logic into separate class
